# Fix slow performance from recursive `__repr__` and `__hash__` for Go (Cherry-pick of #13492)