`BuildGoPackageRequest` is a recursive data structure (DAG). The default `__hash__` and `__repr__` were doing recursive transitive graph walks, which showed up in profiles as pathologically slow, where it looked like building https://github.com/toolchainlabs/external-dns was hung when running `./pants check source:`.

## Benchmark

Before:

```
❯ hyperfine -r 5 './pants_from_sources --no-process-execution-local-cache --no-pantsd check ::'
  Time (mean ± σ):     20.316 s ±  0.877 s    [User: 20.548 s, System: 9.075 s]
  Range (min … max):   19.663 s … 21.668 s    5 runs
```

After:

```
❯ hyperfine -r 5 './pants_from_sources --no-process-execution-local-cache --no-pantsd check ::'
  Time (mean ± σ):     16.163 s ±  0.198 s    [User: 16.970 s, System: 9.193 s]
  Range (min … max):   15.912 s … 16.355 s    5 runs
```

[ci skip-rust]
[ci skip-build-wheels]